Refactor choices in preparation for some new features
choice.scm has been eliminated; the necessary functionality it provided
was folded into rules.scm, and the rest was removed.

generator.scm no longer assumes knowledge about what vector it should
pull phrases from.  This will be used in the future to handle different
generation plans (for example, generating a password that must have a
digit and a special symbol in a secure way).

mantra was adjusted for the new function signatures in generator.scm.