New features for generating passwords that meet complexity requirements.
Password generation now goes through the following steps:
- First, the command line options are used to make a "generation plan."
- Then, the generation plan is "compiled" into a list of phrase vectors;
  for each one, a random phrase is selected, and all the phrases are
  concatenated into the final password.  This process is repeated N times,
  where N is the password count chosen by the user.

The generation plan contains information about which phrase vectors
should actually be used.  Since we now have separate phrase vectors for
phrases that have letters and phrases that have digits and symbols, we
can now, for example, generate a password that definitely has at least
1 digit and 1 symbol.  (Note: The process of "compiling" a generation
plan includes shuffling the phrases, so the randomness of the password
structure is preserved as much as possible; in fact, since we account
for the entropy added by permuting the password phrase structure, in
many realistic scenarios the password entropy is actually greater.)

Code for creating and processing generation plans is found in the new
source file plans.scm.

Other changes:

- rules.scm: Added and exported letter-phrases and digit-symbol-phrases
- generator.scm: random-password was moved to the new plans.scm file;
  additionally, log2 was needed by plans.scm, so it is now exported.
- mantra: --count changed to --password-count to avoid shadowing a
  standard function
- mantra: --min-letter-phrases and --min-digit-symbol-phrases added
  and implemented
- mantra: --print-entropy added; when enabled, prints the calculated
  entropy of each generated password
- mantra: version bumped to 1.1.0