Added a new feature for customizing the set of allowed symbols
The new commandline option -s/--allowed-symbols will allow you to specify
a different set of allowed symbols.  An empty value is allowed, and will
ensure no digit-symbol phrases are generated; if this is provided along
with a nonzero value for min-digit-symbol-phrases, an error is printed.

To accomplish this, phrase vectors are not generated from the rules in
rules.scm until allowed-symbols is set in command line
initialization.  The main print-passwords loop starts by initializing
these phrase vectors, then proceeding to plan generation as before.

Version bumped to 1.2.0.